requirements.txt
asgiref==3.2.3
boto==2.49.0
boto3==1.14.6
botocore==1.17.6
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.41.0
django-environ==0.4.5
django-markdownx==3.0.1
django-ses==1.0.0
django-taggit==1.3.0
docutils==0.15.2
future==0.18.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
jmespath==0.10.0
Markdown==3.2.2
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.12.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
